+++ lib/src/util/functions.dart
@@ -1,7 +1,7 @@
 import '../../cubes.dart';
 
 /// Method util uses to return anything for conditional.
-T? genericCondition<T>({
+T? genericConditional<T>({
   required bool condition,
   required T match,
   T? notMatch,